Abstract static based on methods from directory
