+++ client/src/pages/Directory.tsx
@@ -13,7 +13,14 @@ import Sidebar from '../components/Sidebar';
 import UserModal from '../components/UserModal/UserModal';
 import { getUsers, isError } from '../utils/apiWrapper';
 import { pages } from '../utils/enums';
-import { parseArrayToSemanticDropdownOptions } from '../utils/helpers';
+import {
+  parseArrayToSemanticDropdownOptions,
+  parseSemanticMultiSelectTypes,
+  sortUsersByDate,
+  filterUsersByInterests,
+  filterUsersByRole,
+  filterUsersByTeams,
+} from '../utils/helpers';
 import SSW from '../assets/ssw-form-header.png';
 import DefaultProfile from '../assets/default_profile.png';
 import '../css/Directory.css';
@@ -339,31 +346,6 @@ const Directory = (): ReactElement => {
   };
 
   /**
-   * Parses a semantic multi select into the filter key usable types.
-   *
-   * @param value the value from the multiselect
-   * @returns a conversion from semantic's type to string | string[]
-   */
-  const parseSemanticMultiSelectTypes = (
-    value: string | number | boolean | (string | number | boolean)[],
-  ): string | string[] => {
-    if (typeof value === 'string') {
-      return value;
-    } else if (Array.isArray(value)) {
-      const parsedArray: string[] = [];
-
-      value.forEach((element) => {
-        parsedArray.push(`${element}`);
-      });
-
-      return parsedArray;
-    }
-
-    console.error('Invalid datatype to convert');
-    return [];
-  };
-
-  /**
    * Determines if the filter state is empty.
    *
    * @returns true if filter keys are empty, else false
@@ -373,108 +355,6 @@ const Directory = (): ReactElement => {
     filterKeys.date === '' &&
     filterKeys.interests.length === 0 &&
     filterKeys.teams.length === 0;
-
-  /**
-   * Sorts all of the users in the directory by date.
-   *
-   * @param isAscending the filter string representation, Earliest to Latest, Latest to Earliest, [empty]
-   */
-  const sortUsersByDate = (
-    isAscending: boolean,
-    directory: IUser[],
-  ): IUser[] => {
-    const sortedDirectory: IUser[] = [...directory];
-
-    sortedDirectory.sort((first: IUser, second: IUser): number => {
-      const firstUserDate: Date = new Date(first.dateJoined);
-      const secondUserDate: Date = new Date(second.dateJoined);
-
-      if (firstUserDate > secondUserDate) {
-        // -1 is a magic number
-        return isAscending ? 1 : 0 - 1;
-      } else if (firstUserDate < secondUserDate) {
-        return isAscending ? 0 - 1 : 1;
-      }
-      return 0;
-    });
-
-    return sortedDirectory;
-  };
-
-  /**
-   * Filters out the users that have the specified role.
-   *
-   * @param role the role to filter by
-   */
-  const filterUsersByRole = (role: string, directory: IUser[]): IUser[] => {
-    if (role === '') {
-      return directory;
-    }
-    let filteredDirectory: IUser[] = [...directory];
-
-    filteredDirectory = filteredDirectory.filter(
-      (result: IUser) => result.role === role.toUpperCase(),
-    );
-
-    return filteredDirectory;
-  };
-
-  /**
-   * Filters out the users that have any of the selected interests.
-   *
-   * @param interests the interests to filter by
-   */
-  const filterUsersByInterests = (
-    interests: string[],
-    directory: IUser[],
-  ): IUser[] => {
-    if (typeof interests !== 'object' || interests.length === 0) {
-      return directory;
-    }
-
-    let filteredDirectory: IUser[] = [...directory];
-
-    filteredDirectory = filteredDirectory.filter((user: IUser) => {
-      let hasInterest = true;
-
-      interests.forEach((interest) => {
-        if (!user.interests.includes(interest.toString().toUpperCase())) {
-          hasInterest = false;
-        }
-      });
-
-      return hasInterest;
-    });
-
-    return filteredDirectory;
-  };
-
-  /**
-   * Filters out the users that have any of the selected teams.
-   *
-   * @param interests the interests to filter by
-   */
-  const filterUsersByTeams = (teams: string[], directory: IUser[]): IUser[] => {
-    if (typeof teams !== 'object' || teams.length === 0) {
-      return directory;
-    }
-
-    let filteredDirectory: IUser[] = [...directory];
-
-    filteredDirectory = filteredDirectory.filter((user: IUser) => {
-      let hasTeam = true;
-
-      teams.forEach((team) => {
-        if (!user.currentTeams.includes(team.toUpperCase())) {
-          hasTeam = false;
-        }
-      });
-
-      return hasTeam;
-    });
-
-    return filteredDirectory;
-  };
 
   /**
    * Opens the contributor modal for a specific user.
